Fix version check
