+++ lightly/api/api_workflow_client.py
@@ -20,10 +20,9 @@ from lightly.api.api_workflow_selection import _SelectionMixin
 from lightly.api.api_workflow_upload_dataset import _UploadDatasetMixin
 from lightly.api.api_workflow_upload_embeddings import _UploadEmbeddingsMixin
 from lightly.api.api_workflow_upload_metadata import _UploadCustomMetadataMixin
-from lightly.api.bitmask import BitMask
-from lightly.api.utils import DatasourceType, get_signed_url_destination, getenv
-from lightly.api.version_checking import get_minimum_compatible_version, \
-    version_compare
+from lightly.api.utils import DatasourceType, get_signed_url_destination, get_api_client_configuration
+from lightly.api.version_checking import get_minimum_compatible_version
+from lightly.utils import version_compare
 from lightly.openapi_generated.swagger_client.api.collaboration_api import CollaborationApi
 from lightly.openapi_generated.swagger_client import ScoresApi, QuotaApi, MetaDataConfigurationsApi, PredictionsApi
 from lightly.openapi_generated.swagger_client.api.datasets_api import \
@@ -41,8 +40,6 @@ from lightly.openapi_generated.swagger_client.api.samplings_api import \
     SamplingsApi
 from lightly.openapi_generated.swagger_client.api.tags_api import TagsApi
 from lightly.openapi_generated.swagger_client.api_client import ApiClient
-from lightly.openapi_generated.swagger_client.configuration import \
-    Configuration
 from lightly.openapi_generated.swagger_client.models.dataset_data import \
     DatasetData
 from lightly.utils.reordering import sort_items_by_keys
@@ -120,9 +117,9 @@ class ApiWorkflowClient(_UploadEmbeddingsMixin,
     def check_version_compatibility(self):
         minimum_version = get_minimum_compatible_version()
         if version_compare(__version__, minimum_version) < 0:
-            raise ValueError(f"Incompatible Version of lightly pip package. "
-                             f"Please upgrade to at least version {minimum_version} "
-                             f"to be able to access the api and webapp")
+            raise ValueError(f"Incompatible version of lightly pip package. "
+                             f"Please upgrade to at least version '{minimum_version}' "
+                             f"to be able to access the api.")
 
     @property
     def dataset_id(self) -> str:
@@ -293,26 +290,3 @@ def set_api_client_request_timeout(
         return request_fn(*args, **kwargs)
 
     client.rest_client.request = new_request_fn
-
-
-def get_api_client_configuration(
-    token: Optional[str] = None,
-    raise_if_no_token_specified: bool = True,
-) -> Configuration:
-
-    host = getenv("LIGHTLY_SERVER_LOCATION", "https://api.lightly.ai")
-    ssl_ca_cert = getenv("LIGHTLY_CA_CERTS", None)
-
-    if token is None:
-        token = getenv("LIGHTLY_TOKEN", None)
-    if token is None and raise_if_no_token_specified:
-        raise ValueError(
-            "Either provide a 'token' argument or export a LIGHTLY_TOKEN environment variable"
-        )
-
-    configuration = Configuration()
-    configuration.api_key = {"token": token}
-    configuration.ssl_ca_cert = ssl_ca_cert
-    configuration.host = host
-
-    return configuration
